# swagger.yaml
openapi: 3.0.0
info:
  title: Learn Swagger with NodeJS - Ankush Thakur
  description: Let's learn swagger with NodeJS
  version: 1.0.1

# servers:
#   - url: http://localhost:8000/api/v1
#     description: Local Host v1
#   - url: https://localhost:8000/api/v1
#     description: Local Host v1 - Secure with HTTPS

servers:
  - url: "{protocol}://localhost:8000/api/{version}"
    description: local host {version}
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http

# For Token and Authorization headers
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /lco:
    get:
      tags:
        - String
      summary: Returns a gretting From LCO
      description: A String message from LCO
      responses:
        200:
          description: All Good - Success
          content:
            application/json:
              schema:
                type: string
                example: "myString"
        400:
          description: All Bad - Bad Request
        500:
          description: Internal Server Error

  /lcoobject:
    get:
      tags:
        - Object
      summary: Returns a User Object from LCO
      description: A unique user
      responses:
        200:
          description: All Good - Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  email:
                    type: string
                  number:
                    type: number
        400:
          description: All Bad - Bad Request
        500:
          description: Internal Server Error

  /courses:
    get:
      tags:
        - Array
      summary: Returns all courses
      description: All courses will be shown to user
      responses:
        200:
          description: All Good - Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        400:
          description: All Bad - Bad Request
        500:
          description: Internal Server Error

  /ankush:
    get:
      tags:
        - Swagger Assignment
      summary: Returns array of objects
      description: get all the objects and arrays
      responses:
        200:
          description: All Good - Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  project:
                    type: array
                    items:
                      type: string
                  report:
                    type: array
                    items:
                      type: object
                      properties:
                        grading:
                          type: string
                        frontEnd:
                          type: number
                        backEnd:
                          type: array
                          items:
                            type: object
                            properties:
                              express:
                                type: string
                              node:
                                type: string
                        hobbies:
                          type: array
                          items:
                            type: string
        400:
          description: All Bad - Bad Request
        500:
          description: Internal Server Error